requirements.txt
streamlit
tensorflow
Pillow
numpy
matplotlib
tensorflow_datasets